* Move ColorProviders to their own files * Make default color overrides a little more flexible in uss * A little cleanup
