+++ com.unity.shadergraph/Editor/Drawing/Colors/CategoryColors.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93e3298dd3de748f18fe0a2c486c15f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
