+++ com.unity.shadergraph/Editor/Drawing/Colors/IColorProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8fede14600fd343d592ea0747a5fd164
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
