+++ com.unity.shadergraph/Editor/Drawing/Colors/NoColors.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a22ff676828b743cc8de233f626453ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
