+++ com.unity.shadergraph/Editor/Drawing/Colors/UserColors.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bca71e015fbd848c39feff083689afca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
