+++ com.unity.shadergraph/Editor/Resources/Styles/ColorMode.uss.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 09c2e3747a5e241e3a5c9250fa496b53
+guid: d9a970c4d408c408b888ee0caa8403eb
 ScriptedImporter:
   internalIDToNameTable: []
   externalObjects: {}
