--- com.unity.shadergraph/Editor/Resources/Styles/NodeColors.uss
@@ -1,27 +1,0 @@
-MaterialNodeView #title.Artistic {
-    border-color:#008080;
-}
-
-MaterialNodeView #title.Channel {
-    border-color:#667328;
-}
-
-MaterialNodeView #title.Input {
-    border-color:#99291F;
-}
-
-MaterialNodeView #title.Math {
-    border-color:#1F4799;
-}
-
-MaterialNodeView #title.Procedural {
-    border-color:#993D88;
-}
-
-MaterialNodeView #title.Utility {
-    border-color:#6B83B3;
-}
-
-MaterialNodeView #title.UV {
-    border-color:#268044;
-}
